// frontend/src/api/model/importRegistryView.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * backend | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ImportConcept } from './importConcept';
import type { RegistryCastEntry } from './registryCastEntry';

export interface ImportRegistryView {
    cast?: RegistryCastEntry[];
    concepts?: ImportConcept[];
}